rename Container to SwapApp
