+++ src/App.tsx
@@ -10,7 +10,7 @@ import dark from './components/themes/dark'
 
 import { useStore } from './store';
 import Api from './api';
-import Container from './components/Container'
+import SwapApp from './components/SwapApp'
 
 import Home from './components/Home'
 import About from './components/About'
@@ -171,11 +171,8 @@ const App = () => {
     <BrowserRouter>
       <ThemeProvider theme={false ? dark : light}>
       <Switch>
-          {/* <Container/> refers to the swap app */}
           <Route path="/app">
-            
-          <Container {...{ refresh, connectMetamask, connectWalletConnector }} />
-            
+            <SwapApp {...{ refresh, connectMetamask, connectWalletConnector }} />
           </Route>
           <Route path="/about">
             <About />
